super system/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.3
flask==3.0.0
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3